% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsi.R
\name{rsi}
\alias{rsi}
\title{A function to compute the relative strength index}
\usage{
rsi(input, n)
}
\arguments{
\item{input}{Input time series}

\item{n}{length of RSI}
}
\description{
A function to compute the relative strength index
}
